🚚 Folder restructure (#16)
---------
