+++ .github/workflows/evaluate.yaml
@@ -4,7 +4,7 @@ on:
   pull_request:
     types: [opened, synchronize]
     paths:
-      - "eval/packages/submission-template/**"
+      - "src/eval/submission/**"
   workflow_dispatch: # manually trigger the workflow
 
 jobs:
@@ -38,15 +38,13 @@ jobs:
       - name: Run evaluation
         timeout-minutes: 300 # 5 hours timeout
         run: |
-          cd eval
           GIT_LFS_SKIP_SMUDGE=1 uv sync
           uv run eval --dataset_path=/home/user/datasets/mvtec_loco
-          cd ..
       - name: Upload evaluation results
         uses: actions/upload-artifact@v4
         with:
           name: evaluation-results
-          path: eval/results.json
+          path: results.json
           retention-days: 30
       - name: Add results comment
         uses: actions/github-script@v7
@@ -56,12 +54,12 @@ jobs:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           script: |
             const fs = require('fs');
-            const result = JSON.parse(fs.readFileSync('eval/results.json', 'utf8'));
+            const result = JSON.parse(fs.readFileSync('results.json', 'utf8'));
             let result_string = "";
             for (const [key, value] of Object.entries(result)) {
               result_string += `${key}: ${value}\n`;
             }
-            const raw_metrics = fs.readFileSync('eval/metrics.csv', 'utf8');
+            const raw_metrics = fs.readFileSync('metrics.csv', 'utf8');
             raw_metrics_string = "||seed|k_shot|category|image_score|pixel_score|\n";
             raw_metrics_string += "|---|---|---|---|---|---|\n";
             raw_metrics.split('\n').forEach(line => {
