--- eval/packages/submission-template/README.md
@@ -1,5 +1,0 @@
-# Submission
-
-1. Fork the project
-2. Edit the files in this folder
-3. Create a pull request
