+++ pyproject.toml
@@ -5,17 +5,14 @@ description = "VAND 2025 evaluation script"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
-    "submission-template",
     "anomalib[core]",
     "scikit-learn",
+    # TODO: Add your additional dependencies here
 ]
 
 [tool.uv.sources]
-submission-template = { workspace = true }
 anomalib = { git = "https://github.com/openvinotoolkit/anomalib.git", branch = "main" }
 
-[tool.uv.workspace]
-members = ["packages/*"]
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/eval"]
@@ -26,3 +23,12 @@ eval = "eval.evaluate:eval"
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.ruff]
+line-length = 120
+
+[tool.ruff.lint]
+select = ["E", "F", "W", "I"]
+
+[tool.ruff.format]
+# Optional: Configure Ruff's formatter
