+++ src/eval/submission/__init__.py
@@ -1,0 +1,1 @@
+"""VAND 2025 Challenge Submission Template."""
